confirm_cmdの追加
